requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
Django==4.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.4.0
python-barcode==0.14.0
sqlparse==0.4.4
tomli==2.0.1
